Update README.md
As the radius of two atoms increase, the potential energy between atoms decreases as they approach each other. After the pairwise_energy past the zero and reached the lowest pairwise_energy the pairwise_energy of the two atoms increase and keep approaching zero. E_list shows how the pairwise_energy change as the radius of the atoms change.